refactor code.

+++ app/src/main/java/ru/firstset/whereisuser/util/UtilsOther.java
@@ -13,7 +13,7 @@ public class UtilsOther {
         return sdf.format(new Date());
     }
 
-    public static boolean compareLocations(Float lastKnownLatitude, Float lastKnownLongitude, Float currentLatitude, Float currentLongitude) {
+    public static boolean compareLocations(Double lastKnownLatitude, Double lastKnownLongitude, Double currentLatitude, Double currentLongitude) {
         if ((lastKnownLatitude.compareTo(currentLatitude) == 0) &
                 (lastKnownLongitude.compareTo(currentLongitude) == 0)) {
             Log.v("compareLocations","true");
@@ -27,4 +27,13 @@ public class UtilsOther {
             return false;
         }
     }
+//    public static float getFloatRoun3(Double doubleIn){
+////        String result = String.format("%.3f",doubleIn);
+////        return Float.valueOf(result);
+////        double value = 34.777774;
+////        double scale = Math.pow(10, 3);
+////        Float result = Math.ceil(value * scale) / scale;
+//        float result = (float) ((float) Math.round(doubleIn * 100000d) / 100000d);
+//        return result;
+//    }
 }
